requirements.txt
pyTelegramBotAPI>=4.20.0
google-genai>=1.0.0
python-dotenv>=1.0.1
finnhub-python>=2.4.20
yfinance>=0.2.40
requests>=2.32.0
pandas>=2.2.0
pytz>=2024.1
psutil>=5.9.0
